Refactor Wasp.LSP.State into multiple files
